Use Conan for managing dependencies (#338)
* Use Conan for dependencies

* Switch to Conan

* Add conan files for a few libraries
Move python scripts to python 3

* Added IU_ENABLE_FFMPEG option
Fixed MediaInfo's report i18n
Fixed python scripts in Utils directory
Fixed package recipe for mega.nz SDK

* Fixed ffmpeg usage in AboutDlg and CMakeLists.txt

* Update README and COMPILING.MD

* Fixed COMPILING.MD

* Remove remotes from conan_cmake_install in CMakeLists.txt

* Fix in CMakeLists.txt

* Fix comilation on linux

* Fixes in CMakeLists.txt

* Fix CMakeLists curl dependency

* Update build instruction

+++ Source/Core/DownloadTaskTest.cpp
@@ -65,7 +65,7 @@ TEST_F(DownloadTaskTest, Download)
         EXPECT_TRUE(ok);
         EXPECT_EQ(url, item.url);
         EXPECT_EQ("test.png", item.displayName);
-        EXPECT_EQ(123, reinterpret_cast<int>(item.id));
+        EXPECT_EQ(123, reinterpret_cast<intptr_t>(item.id));
         EXPECT_EQ(referer, item.referer);
         EXPECT_EQ(200, responseCode);
         EXPECT_FALSE(item.fileName.empty());
